move shared graph filter code to seperate file
