+++ weblab/stories/admin.py
@@ -1,13 +1,8 @@
 from django.contrib import admin
 
-from .models import Story, StoryText, StoryGraph
+from .models import Story, StoryGraph, StoryText
 
 
 admin.site.register(Story)
 admin.site.register(StoryText)
 admin.site.register(StoryGraph)
-
-
-
-
-
